Make another walcraft test more robust
There were two issues with the test at page boundaries:

1. If the first logical message with 10 bytes payload crossed a page
   boundary, the calculated 'base_size' was too large because it included
   the page header.

2. If it was inserted near the end of a page so that there was not
   enough room for another one, we did "remaining_lsn += XLOG_BLCKSZ"
   but that didn't take into account the page headers either.

As a result, the test would fail if the WAL insert position at the
beginning of the test was too close to the end of a WAL page. Fix the
calculations by repeating the 10-byte logical message if the starting
position is not suitable.

I bumped into this with PR #7377; it changed the arguments of a few
SQL functions in neon_test_utils extension, which changed the WAL
positions slightly, and caused a test failure.